fix: Address compiler warnings in modified files
- Fix sign comparison warning in sqlite_duckdb_vfs_cache.cpp
- Fix bug where wrong operator was passed to children in sqlite_insert.cpp
  (was using original plan instead of casted plan when types need conversion)

These issues were present in the main branch but fixed since we're modifying the files.

+++ src/sqlite_duckdb_vfs_cache.cpp
@@ -388,7 +388,7 @@ int SQLiteDuckDBCacheVFS::Open(sqlite3_vfs *vfs, const char *filename, sqlite3_f
 
 	try {
 		// Ensure SQLite allocated enough space for our file structure
-		if (vfs->szOsFile < sizeof(SQLiteDuckDBCachedFile)) {
+		if (vfs->szOsFile < static_cast<int>(sizeof(SQLiteDuckDBCachedFile))) {
 			return SQLITE_CANTOPEN;
 		}
